ext_symlink.sh : use timeout to wait for pty This commit restores the below functionality
symlink: Run symlinks 25 times when exec
b69395e9bc645ab209b52bacf85cd9aeb41ba372 @corphish corphish committed on Nov 3, 2016

But uses timeout instead of running it 25 times.

+++ proprietary/bin/ext_symlink.sh
@@ -3,12 +3,22 @@
 term="/dev/pts/* "
 
 if [ "$1" = "-p" ]; then
-	link=`getprop sys.symlink.pty`
-	rm ${link##${term}}
-	ln -s $link;
-	setprop sys.symlink.notify 0
+	timeout=25
+	while [ "$timeout" -ge 0 ] ; do 
+		link=`getprop sys.symlink.pty`
+		if [ -f ${link##${term}} ]; then
+			rm ${link##${term}}
+			ln -s $link;
+		    setprop sys.symlink.notify 0
+		  break;
+		fi
+		sleep 1
+		timeout=$((timeout - 1))
+	done
 elif [ "$1" = "-u" ]; then
 	link=`getprop sys.symlink.umts_router`
-	rm ${link##${term}}
+	if [ -f ${link##${term}} ]; then
+		rm ${link##${term}}
+	fi
 	ln -s $link;
 fi
